Move Economy and EconomyType to API module
