+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/rewardtypes/external/MoneyRewardType.java
@@ -1,9 +1,9 @@
 package com.oheers.fish.competition.rewardtypes.external;
 
-import com.oheers.fish.Economy;
+import com.oheers.fish.api.economy.Economy;
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.api.reward.RewardType;
-import com.oheers.fish.economy.EconomyType;
+import com.oheers.fish.api.economy.EconomyType;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -13,7 +13,7 @@ public class MoneyRewardType implements RewardType {
 
     @Override
     public void doReward(@NotNull Player player, @NotNull String key, @NotNull String value, Location hookLocation) {
-        Economy economy = EvenMoreFish.getInstance().getEconomy();
+        Economy economy = Economy.getInstance();
         int amount;
         try {
             amount = Integer.parseInt(value);
