+++ even-more-fish-plugin/src/main/java/com/oheers/fish/selling/SellHelper.java
@@ -1,7 +1,7 @@
 package com.oheers.fish.selling;
 
 import com.devskiller.friendly_id.FriendlyId;
-import com.oheers.fish.Economy;
+import com.oheers.fish.api.economy.Economy;
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import com.oheers.fish.config.MainConfig;
@@ -23,8 +23,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.sql.Timestamp;
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
 import java.time.Instant;
 import java.util.*;
 
@@ -32,7 +30,7 @@ public class SellHelper {
 
     private final Inventory inventory;
     private final Player player;
-    
+
     private int fishCount;
 
     public static void sellInventoryGui(@NotNull InventoryGui gui, @NotNull HumanEntity humanEntity) {
@@ -69,8 +67,8 @@ public class SellHelper {
             }
         }
 
-        Economy economy = EvenMoreFish.getInstance().getEconomy();
-        if (economy != null && economy.isEnabled()) {
+        Economy economy = Economy.getInstance();
+        if (economy.isEnabled()) {
             economy.deposit(this.player, totalWorth, true);
         }
 
@@ -81,7 +79,7 @@ public class SellHelper {
         // sending the sell message to the player
 
         Message message = new Message(ConfigMessage.FISH_SALE);
-        if (economy == null || !economy.isEnabled()) {
+        if (!economy.isEnabled()) {
             message.setSellPrice("0");
         } else {
             message.setSellPrice(economy.getWorthFormat(sellPrice, true));
